Feat/mission alarm

+++ src/main/java/com/moing/backend/domain/missionArchive/application/service/MissionArchiveDeleteUseCase.java
@@ -58,15 +58,18 @@ public class MissionArchiveDeleteUseCase {
         Mission mission = missionQueryService.findMissionById(missionId);
         Team team = mission.getTeam();
 
-        // 사진 제출 했다면, s3 삭제 로직
-        if (mission.getWay() == MissionWay.PHOTO && missionArchiveQueryService.isDone(memberId, missionId)) {
+        MissionArchive deleteArchive = missionArchiveQueryService.findOneMyArchive(memberId, missionId,count).get(0);
+        MissionState missionState = missionStateQueryService.findMissionState(member, mission);
 
+        LocalDateTime createdDate = deleteArchive.getCreatedDate();
+        LocalDateTime today = LocalDateTime.now();
+
+        // 반복미션이면서 오늘 이전에 한 인증은 인증 취소할 수 없도록
+        if (mission.getType().equals(MissionType.REPEAT) && createdDate.toLocalDate().isBefore(today.toLocalDate())) {
+            throw new NoAccessMissionArchiveException();
         }
 
-        MissionArchive deleteArchive = missionArchiveQueryService.findOneMyArchive(memberId, missionId,count).get(0);
         missionArchiveDeleteService.deleteMissionArchive(deleteArchive);
-
-        MissionState missionState = missionStateQueryService.findMissionState(member, mission);
         missionStateDeleteService.deleteMissionState(missionState);
 
         return deleteArchive.getId();
